feat: smooth overlap resolution
