+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "la_bonne_echappee",
-  "version": "1.0.23",
+  "version": "1.0.24",
   "description": "",
   "main": "simulation.js",
   "scripts": {
-    "test": "node test/relayCluster.test.js && node test/relayCycle.test.js && node test/eventBus.test.js",
+    "test": "node test/relayCluster.test.js && node test/relayCycle.test.js && node test/eventBus.test.js && node test/overlapResolution.test.js",
     "lint": "eslint .",
     "version": "echo version script removed"
   },
